Use a specific name for SCC
There are some tests [1] that fail because the SCC name is too large.
The SCC name when the aot test flag is set is derived from the test
name. In order to prevent this issue, simply use the same name for all
tests. This is ok because prior to and after running tests, all SCCs are
destroyed. Additionally, this does not impact any future ability to run
tests in parallel as this is already not possible; the fact that all
SCCs are destroyed between tests means that test could impact the
validity of another.

[1] https://github.com/eclipse-openj9/openj9/issues/14461

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>